tweak(ui): Deconstruct Graph components
- Decouple filter bar into it's own component
- Decouple graph header
