+++ client/components/FilterBar/FilterBar.scss
@@ -1,55 +1,7 @@
-%title {
-  height: 40px;
-  margin: 0;
-  color: #555;
-  font-size: 32px;
-  line-height: 40px;
-}
-
-.container {
-  position: relative;
-  grid-area: graph;
-  height: 496px;
-  padding: 24px;
-  background: #fff;
-}
-
-.balance {
-  @extend %title;
-  font-weight: normal;
-}
-
-.name {
-  @extend %title;
-  font-weight: bold;
-}
-
-.returns {
-  top: 65px;
-  height: 20px;
-  margin: 0;
-  color: #888;
-  font-size: 14px;
-  font-weight: normal;
-  line-height: 20px;
-
-  &--pos {
-    color: #21ce99;
-  }
-
-  &--neg {
-    color: #f45431;
-  }
-
-  .date {
-    color: #888;
-  }
-}
-
 .filter {
   display: flex;
   height: 32px;
-  padding-top: 19px;
+  padding-top: 12px;
   border-bottom: 1px solid #999;
   justify-content: space-between;
 }
