+++ client/components/FilterBar/models/FilterBar.ts
@@ -1,0 +1,6 @@
+import { PortfolioParam } from '@/store/models/portfolio';
+
+export interface FilterBarProps {
+  filter: PortfolioParam;
+  filterClick(filter: PortfolioParam): void;
+}
