+++ client/components/Graph/models/Graph.ts
@@ -4,9 +4,8 @@ export interface GraphProps {
   data: PortfolioData[];
   filter: PortfolioParam;
   height: number;
-  name: string;
   width: number;
-  filterClick(filter: PortfolioParam): void;
+  setNext(next: PortfolioData, newDate?: string): void;
 }
 
 export interface StateProps {
