+++ client/pages/Portfolio/Overview/Overview.scss
@@ -1,0 +1,43 @@
+%title {
+  height: 40px;
+  margin: 0;
+  color: #555;
+  font-size: 32px;
+  line-height: 40px;
+}
+
+.overview {
+  padding-bottom: 12px;
+}
+
+.balance {
+  @extend %title;
+  font-weight: normal;
+}
+
+.name {
+  @extend %title;
+  font-weight: bold;
+}
+
+.returns {
+  top: 65px;
+  height: 20px;
+  margin: 0;
+  color: #888;
+  font-size: 14px;
+  font-weight: normal;
+  line-height: 20px;
+
+  &--pos {
+    color: #21ce99;
+  }
+
+  &--neg {
+    color: #f45431;
+  }
+
+  .date {
+    color: #888;
+  }
+}
