+++ client/pages/Portfolio/Overview/models/Overview.ts
@@ -1,0 +1,7 @@
+export interface OverviewProps {
+  name: string;
+  balance: number;
+  returns: number;
+  percentage: number;
+  date: string;
+}
