+++ client/pages/Portfolio/Portfolio.tsx
@@ -2,20 +2,25 @@ import React, { FC, useEffect, useState } from 'react';
 import { Dispatch } from 'redux';
 import { connect } from 'react-redux';
 import { AppState } from '@/store/models/store';
-import { PortfolioParam } from '@/store/models/portfolio';
+import { PortfolioParam, PortfolioData } from '@/store/models/portfolio';
 import { userDataAction } from '@/store/actions/user';
 import { portfolioListAction, selectedPortfolioAction } from '@/store/actions/portfolio';
 import { analyticsTotalAction, analyticsAnnualizeAction, analyticsPastYearAction } from '@/store/actions/analytics';
 import * as _ from '@/utils/collection-util';
 import Sidebar from '@/components/Sidebar/Sidebar';
-import Graph from '@/components/Graph/PortfolioGraphContainer';
+import Graph from '@/components/Graph/Graph';
 import Analytics from '@/components/Analytics/Analytics';
-import styles from './Portfolio.scss';
+import FilterBar from '@/components/FilterBar/FilterBar';
 import { PortfolioProps, StateProps, DispatchProps } from './models/Portfolio';
+import styles from './Portfolio.scss';
+import { convertToMoney, convertToPercent } from '@/utils/util';
+import Overview from './Overview/Overview';
 
 const Portfolio: FC<PortfolioProps> = ({
   id,
   portfolios,
+  name,
+  data,
   onInit,
   getPortfolioData,
   getAnalyticsData,
@@ -24,6 +29,19 @@ const Portfolio: FC<PortfolioProps> = ({
     range: '180D',
     data: 'Cumulative Returns',
   });
+  const [balance, setBalance] = useState<number>(0);
+  const [returns, setReturns] = useState<number>(0);
+  const [percentage, setPercentage] = useState<number>(0);
+  const [date, setDate] = useState<string>('');
+
+  const setNext = (next: PortfolioData, newDate: string = '') => {
+    const start = data[0];
+    const nextReturns = next.cumulativeReturns - start.cumulativeReturns;
+    setBalance(next.balance);
+    setReturns(nextReturns);
+    setPercentage(nextReturns / start.balance * 100);
+    setDate(newDate);
+  };
 
   useEffect(onInit, []);
 
@@ -38,12 +56,23 @@ const Portfolio: FC<PortfolioProps> = ({
   return (
     <div className={styles.portfolio}>
       <Sidebar portfolioClick={getPortfolioData} />
-      <Graph
-        height={340}
-        width={1064}
-        filter={filter}
-        filterClick={setFilter}
-      />
+      <div className={styles.graph}>
+        <Overview
+          name={name}
+          balance={balance}
+          returns={returns}
+          percentage={percentage}
+          date={date}
+        />
+        <Graph
+          height={340}
+          width={1064}
+          filter={filter}
+          data={data}
+          setNext={setNext}
+        />
+        <FilterBar filter={filter} filterClick={setFilter} />
+      </div>
       <Analytics />
     </div>
   );
@@ -52,6 +81,8 @@ const Portfolio: FC<PortfolioProps> = ({
 const mapStateToProps = (state: AppState): StateProps => ({
   portfolios: state.portfolio.list,
   id: state.portfolio.selected.id,
+  data: state.portfolio.selected.data,
+  name: state.portfolio.selected.name,
 });
 
 const mapDispatchToProps = (dispatch: Dispatch): DispatchProps => ({
