+++ client/pages/Portfolio/models/Portfolio.ts
@@ -1,8 +1,10 @@
-import { IPortfolio, PortfolioParam } from '@/store/models/portfolio';
+import { IPortfolio, PortfolioParam, PortfolioData } from '@/store/models/portfolio';
 
 export interface StateProps {
   portfolios: IPortfolio[];
   id: string;
+  data: PortfolioData[];
+  name: string;
 }
 
 export interface DispatchProps {
